+++ src/main/java/games/twinhead/moreslabsstairsandwalls/block/magma/MagmaStairs.java
@@ -1,0 +1,48 @@
+package games.twinhead.moreslabsstairsandwalls.block.magma;
+
+import net.minecraft.block.*;
+import net.minecraft.enchantment.EnchantmentHelper;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.LivingEntity;
+import net.minecraft.entity.damage.DamageSource;
+import net.minecraft.server.world.ServerWorld;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.world.World;
+import net.minecraft.world.WorldAccess;
+
+import java.util.Random;
+
+@SuppressWarnings("deprecation")
+public class MagmaStairs extends StairsBlock {
+
+    public MagmaStairs(Settings settings) {
+        super(Blocks.MAGMA_BLOCK.getDefaultState(), settings);
+    }
+
+    @Override
+    public void onSteppedOn(World world, BlockPos pos, BlockState state, Entity entity) {
+        if (!entity.bypassesSteppingEffects() && entity instanceof LivingEntity && !EnchantmentHelper.hasFrostWalker((LivingEntity)entity)) {
+            entity.damage(DamageSource.HOT_FLOOR, 1.0f);
+        }
+        super.onSteppedOn(world, pos, state, entity);
+    }
+
+    @Override
+    public void scheduledTick(BlockState state, ServerWorld world, BlockPos pos, Random random) {
+        BubbleColumnBlock.update(world, pos.up(), state);
+    }
+
+    @Override
+    public BlockState getStateForNeighborUpdate(BlockState state, Direction direction, BlockState neighborState, WorldAccess world, BlockPos pos, BlockPos neighborPos) {
+        if (direction == Direction.UP && neighborState.isOf(Blocks.WATER)) {
+            world.createAndScheduleBlockTick(pos, this, 20);
+        }
+        return super.getStateForNeighborUpdate(state, direction, neighborState, world, pos, neighborPos);
+    }
+
+    @Override
+    public void onBlockAdded(BlockState state, World world, BlockPos pos, BlockState oldState, boolean notify) {
+        world.createAndScheduleBlockTick(pos, this, 20);
+    }
+}