+++ src/main/java/games/twinhead/moreslabsstairsandwalls/mixin/SpreadableBlockMixin.java
@@ -1,7 +1,7 @@
 package games.twinhead.moreslabsstairsandwalls.mixin;
 
 import games.twinhead.moreslabsstairsandwalls.block.ModBlocks;
-import games.twinhead.moreslabsstairsandwalls.block.SpreadableSlabBlock;
+import games.twinhead.moreslabsstairsandwalls.block.spreadable.SpreadableSlabBlock;
 import games.twinhead.moreslabsstairsandwalls.registry.ModBlockTags;
 import net.minecraft.block.*;
 import net.minecraft.server.world.ServerWorld;