+++ changelog.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [2.3.0- Fabric 1.18.2]
+
+## Added
+- Dirt path slab, stairs and wall
+    - Use a shovel on any dirt based block to acquire
+- Strippable Log Blocks
+    - Use an axe on any Log/Wood slab, stair or wall to strip it.
+## Changes
+- All blocks now apply slipperiness/velocity multipliers to entities on bottom slabs/stairs
+- Soul sand/Magma Blocks now create bubble columns when below a water source or when waterlogged.
+- Soul sand slabs now have a lower collision shape
+- Soul Speed now works with Soul Sand/Soil slabs stairs and walls
+- Magama slabs, stairs, and walls now damage entities above.
+- Basalt, Melon, Pumpkin, and Hay walls now show proper top/bottom texture
+- Leaves slabs, stairs, and walls can now be mined faster with shears
+- Leaves slabs and stairs can be set if they are persistent or not
+
+## Fixed
+- issue where leaves stairs blocks would turn into a dirt/grass block.
 
 ## [2.2.1 - Fabric 1.18.2]
 ### Fixed